requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Core Airflow
apache-airflow==2.7.3

# Snowflake connector
snowflake-connector-python==3.3.1

# HTTP requests
requests==2.31.0

# Additional utilities
python-dotenv==1.0.0

